Using ONC RPC to connect the two players
The ONC RPC technologie provides an abstraction on how player1 and
player2 are connected and how data is transmitted between them, which
will simplify and organise the code!

From now on, player2 has its own timer to move the ball which will enable
it to update, independently of player2:

 - the ball position
 - the score (P1 and P2)

Therefore, only the rackets x-coordinates will be interchanged over the
network. By the way, this great optimization would not have been possible
without using remote functions LANCHBALLP1 and LANCHBALLP2.

In addition, racket positions are transmitted between the two players
using player1's GETRACK1 remote function. Sockets to transmit data are
created automatically by the RPC mechanism and doesn't need to be created
and configured explicitally on the code.

The use of SDL_PollEvent() (instead of SDL_WaitEvent() blocking function),
will enable the code to poll asynchronously for event in the game loop.
This will not block the program while handling events which will avoid the
use of threads in the code.


Signed-off-by: Issam Abdallah <iabdallah@yandex.com>